requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
Flask-Compress==1.13
pyicloud==1.0.0
WTForms==3.0.1
python-dotenv==1.0.0
Werkzeug<4.0
pytz
geopy
folium
requests
schedule==1.2.0
PyMySQL==1.1.1
mysqlclient==2.2.0 